google-go-lang: update to go1.4.3
This addresses a couple recent CVEs around http smuggling in the
'net/http' standard library.

Additionally, despite requests by some in the community, I'm not yet
upgrading to go1.5.1. This newer version of go is great, no doubt, but
is a full transition of the compiler itself being written in C, to
golang source. So there is a definite bootstrapping situation.

Due to the [lack of] dependency manangement in slackbuilds, there is not
a clean way to keep both compilers. I was hoping the next release of
slackware would include gcc version 5.0 (or newer). This way we could
bootstrap with gccgo.
I'm not sure what I'll do otherwise.